\name{getNChainsPop.twDEMC}
\alias{getNChainsPop.twDEMC}
\title{getNChainsPop twDEMC}
\description{Extracts the number of chains per population}
\usage{
\method{getNChainsPop}{twDEMC}(res, ...)
}
\arguments{
  \item{res}{object of class twDEMC}
  \item{\dots}{
}
}

\value{integer, number of chains per population in twDEMC}

\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}



\seealso{\code{\link{getNGen.twDEMC}}
\code{\link{subChains.twDEMC}}
,\code{\link{twDEMCInt}}}

