\name{getNPops.twDEMC}
\alias{getNPops.twDEMC}
\title{getNPops twDEMC}
\description{Extracts the number of populations}
\usage{
\method{getNPops}{twDEMC}(res, ...)
}
\arguments{
  \item{res}{object of class twDEMC}
  \item{\dots}{
}
}

\value{integer, number of populations in twDEMC}

\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}



\seealso{\code{\link{getNGen.twDEMC}}
,\code{\link{subChains.twDEMC}}
,\code{\link{twDEMCInt}}}

