\name{twDEMC-internal}
\title{Internal twDEMC objects}
\alias{.calcRhat2}
\alias{.calcTemperatedDiffLogDen}
\alias{.colNames}
\alias{.cut2dim2}
\alias{.cut2dim3}
\alias{.detrendMatrix}
\alias{.doDEMCStep}
\alias{.doDEMCSteps}
\alias{.fCheckAlphaSlopeSplit}
\alias{.fDAlphaSlopeI}
\alias{.fPVarI}
\alias{.generateXPropThin}
\alias{.getResFLogDenNames}
\alias{.ofLnormMLE}
\alias{.plotMarginal2D.levelplot}
\alias{.tmp.f}
\alias{.tmp.f.plot.calcDEMCTempDiffLogDen2}
\alias{.twCoefLnormMLE_depr}
\alias{.xStepParallel}
\alias{.xStepSnooker}
\description{Internal twDEMC objects.}
\details{These are not to be called by the user.}
\keyword{internal}